requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
chardet==2.3.0
docopt==0.6.2
docx2txt==0.6
EbookLib==0.15
lxml==4.2.5
miette==1.3
pdfminer.six==20181108
Pillow==5.3.0
pocketsphinx==0.1.3
pycryptodome==3.7.2
python-docx==0.8.7
python-pptx==0.6.5
six==1.10.0
sortedcontainers==2.1.0
SpeechRecognition==3.6.3
textract==1.6.1
xlrd==1.0.0
XlsxWriter==1.1.2
